requirements.txt
Click==7.0
Flask==1.1.1
gunicorn
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug
feedparser
newspaper3k
flask-cors